+++ .vscode/settings.json
@@ -10,5 +10,5 @@
     "source.fixAll": true
   },
   "typescript.preferences.importModuleSpecifier": "non-relative",
-  "cSpell.words": ["appjam", "hookform", "sopkathon", "Sopt"]
+  "cSpell.words": ["appjam", "hookform", "JOINTSEMINAR", "sopkathon", "Sopt", "SOPTERM"]
 }
