+++ components/members/detail/GenLogo.tsx
